requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai>=1.40.0
pdfplumber==0.11.4
jinja2==3.1.4
PyMuPDF==1.24.10
pandas==2.2.2
google-genai
python-dotenv
pydantic
